+++ cluster_agent/agent.py
@@ -1,34 +1,21 @@
-from cluster_agent.settings import SETTINGS, CLUSTER_API_HEADER
-from cluster_agent.utils.request import (
-    async_req,
-    general_slurmrestd_request,
-)
+import asyncio
+
+from cluster_agent.utils.exception import SlurmrestdError
+from cluster_agent.identity.cluster_api import backend_client as cluster_api_client
+from cluster_agent.identity.slurmrestd import backend_client as slurmrestd_client
 
 import hostlist
-import requests
-import json
-
-# [nest-asyncio docs](https://pypi.org/project/nest-asyncio/)
-import nest_asyncio
-
-nest_asyncio.apply()
-
-
-def cluster_api_header():
-
-    return {"Content-Type": "application/json", "Authorization": SETTINGS.API_KEY}
 
 
 async def upsert_partitions():
 
-    partitions = await general_slurmrestd_request("/slurm/v0.0.36/partitions")
+    r = await slurmrestd_client.get("/slurm/v0.0.36/partitions")
+    SlurmrestdError.require_condition(
+        r.status_code == 200, f"Slurmrestd returned {r.status_code} when calling {r.url}: {r.text}"
+    )
+    partitions = r.json()
 
-    # arguments passed to async request handler
-    urls = list()
-    methods = list()
-    params = list()
-    data = list()
-    header = CLUSTER_API_HEADER
+    tasks = list()
 
     for partition in partitions["partitions"]:
 
@@ -42,27 +29,28 @@ async def upsert_partitions():
             "partition": partition,
         }
 
-        urls.append(SETTINGS.BASE_API_URL + f"/agent/partitions/{partition['name']}")
-        methods.append("PUT")
-        params.append(None)
-        data.append(json.dumps(payload))
+        tasks.append(
+            cluster_api_client.put(
+                f"/cluster/agent/partitions/{partition['name']}",
+                json=payload,
+            )
+        )
 
-    responses = await async_req(urls, methods, header, params, data)
+    responses = await asyncio.gather(*tasks)
 
     # return a list containing just the responses' status, e.g. [200, 400]
-    return [response.status for response in responses]
+    return [response.status_code for response in responses]
 
 
 async def upsert_nodes():
 
-    nodes = await general_slurmrestd_request("/slurm/v0.0.36/nodes")
+    r = await slurmrestd_client.get("/slurm/v0.0.36/nodes")
+    SlurmrestdError.require_condition(
+        r.status_code == 200, f"Slurmrestd returned {r.status_code} when calling {r.url}: {r.text}"
+    )
+    nodes = r.json()
 
-    # arguments passed to async request handler
-    urls = list()
-    methods = list()
-    params = list()
-    data = list()
-    header = CLUSTER_API_HEADER
+    tasks = list()
 
     for node in nodes["nodes"]:
 
@@ -72,41 +60,41 @@ async def upsert_nodes():
             "node": node,
         }
 
-        urls.append(SETTINGS.BASE_API_URL + f"/agent/nodes/{node['name']}")
-        methods.append("PUT")
-        params.append(None)
-        data.append(json.dumps(payload))
+        tasks.append(
+            cluster_api_client.put(
+                f"/cluster/agent/nodes/{node['name']}",
+                json=payload,
+            )
+        )
 
-    responses = await async_req(urls, methods, header, params, data)
+    responses = await asyncio.gather(*tasks)
 
     # return a list containing just the responses' status, e.g. [200, 400]
-    return [response.status for response in responses]
+    return [response.status_code for response in responses]
 
 
 async def update_diagnostics():
 
-    diagnostics = await general_slurmrestd_request("/slurm/v0.0.36/diag/")
-
-    response = requests.post(
-        SETTINGS.BASE_API_URL + "/agent/diagnostics",
-        headers=CLUSTER_API_HEADER,
-        data=json.dumps(diagnostics),
+    r = await slurmrestd_client.get("/slurm/v0.0.36/diag/")
+    SlurmrestdError.require_condition(
+        r.status_code == 200, f"Slurmrestd returned {r.status_code} when calling {r.url}: {r.text}"
     )
+    diagnostics = r.json()
 
-    # return a list container the status code response, e.g. [200]
-    return [response.status_code]
+    response = await cluster_api_client.post("/cluster/agent/diagnostics", json=diagnostics)
+
+    return response.status_code
 
 
 async def upsert_jobs():
 
-    jobs = await general_slurmrestd_request("/slurm/v0.0.36/jobs")
+    r = await slurmrestd_client.get("/slurm/v0.0.36/jobs")
+    SlurmrestdError.require_condition(
+        r.status_code == 200, f"Slurmrestd returned {r.status_code} when calling {r.url}: {r.text}"
+    )
+    jobs = r.json()
 
-    # arguments passed to async request handler
-    urls = list()
-    methods = list()
-    params = list()
-    data = list()
-    header = CLUSTER_API_HEADER
+    tasks = list()
 
     for job in jobs["jobs"]:
 
@@ -116,12 +104,14 @@ async def upsert_jobs():
             "job": job,
         }
 
-        urls.append(SETTINGS.BASE_API_URL + f"/agent/jobs/{job['job_id']}")
-        methods.append("PUT")
-        params.append(None)
-        data.append(json.dumps(payload))
+        tasks.append(
+            cluster_api_client.put(
+                f"/cluster/agent/jobs/{job['job_id']}",
+                json=payload,
+            )
+        )
 
-    responses = await async_req(urls, methods, header, params, data)
+    responses = await asyncio.gather(*tasks)
 
     # return a list container the status code response, e.g. [200]
-    return [response.status for response in responses]
+    return [response.status_code for response in responses]