+++ cluster_agent/settings.py
@@ -1,14 +1,14 @@
-from pydantic.error_wrappers import ValidationError
+import sys
+from functools import lru_cache
+from pathlib import Path
+
 from pydantic import BaseSettings, Field
+from pydantic.error_wrappers import ValidationError
 
 from cluster_agent.utils.logging import logger
 
-from functools import lru_cache
-import sys
-
 
 _URL_REGEX = r"http[s]?://.+"
-_API_KEY_REGEX = r"([a-zA-Z0-9])\w+"
 
 
 class Settings(BaseSettings):
@@ -18,9 +18,16 @@ class Settings(BaseSettings):
 
     # cluster api info
     BASE_API_URL: str = Field("https://rats.omnivector.solutions", regex=_URL_REGEX)
-    API_KEY: str = Field("ratsratsrats", regex=_API_KEY_REGEX)
 
     SENTRY_DSN: str = Field("https://rats.sentry.com", regex=_URL_REGEX)
+
+    # Auth0 config for machine-to-machine security
+    AUTH0_DOMAIN: str = Field("omnivector.auth0.com")
+    AUTH0_AUDIENCE: str = Field("https://domain.omnivector.solutions")
+    AUTH0_CLIENT_ID: str = Field("abcde12345")
+    AUTH0_CLIENT_SECRET: str = Field("abcde12345")
+
+    CACHE_DIR = Path.home() / ".cache/cluster-agent"
 
     class Config:
 
@@ -38,8 +45,3 @@ def init_settings() -> Settings:
 
 
 SETTINGS = init_settings()
-
-CLUSTER_API_HEADER = {
-    "Content-Type": "application/json",
-    "Authorization": SETTINGS.API_KEY,
-}